separate out tests as per the database tests
