+++ src/blackboxTests/src/test/kotlin/org/vaccineimpact/api/blackboxTests/BlackboxTestsSuite.kt
@@ -5,6 +5,8 @@ import org.junit.BeforeClass
 import org.junit.runner.RunWith
 import org.junit.runners.Suite
 import org.vaccineimpact.api.blackboxTests.tests.*
+import org.vaccineimpact.api.blackboxTests.tests.BurdenEstimates.RetrieveBurdenEstimateTests
+import org.vaccineimpact.api.blackboxTests.tests.BurdenEstimates.UploadBurdenEstimateTests
 import org.vaccineimpact.api.test_helpers.DatabaseCreationHelper
 
 // Keep these sorted alphabetically, for consistency
@@ -12,7 +14,6 @@ import org.vaccineimpact.api.test_helpers.DatabaseCreationHelper
 @Suite.SuiteClasses(
         AccessLogTests::class,
         AuthenticationTests::class,
-        BurdenEstimateTests::class,
         CoverageTests::class,
         CreateUserTests::class,
         DemographicTests::class,
@@ -23,8 +24,10 @@ import org.vaccineimpact.api.test_helpers.DatabaseCreationHelper
         ModelTests::class,
         PasswordTests::class,
         ResponsibilityTests::class,
+        RetrieveBurdenEstimateTests::class,
         ScenarioTests::class,
         TouchstoneTests::class,
+        UploadBurdenEstimateTests::class,
         UserTests::class
 )
 class BlackboxTestsSuite
